Logs

+++ EmprestimoLivros.API/Controllers/EmployeeController.cs
@@ -1,9 +1,9 @@
 using EmployeesRelation.API.Database;
 using EmployeesRelation.API.Dto;
+using EmployeesRelation.API.Filters;
 using EmployeesRelation.API.Models;
 using Microsoft.AspNetCore.Mvc;
 using System.Net.Mime;
-using System.Text.Json;
 
 namespace EmployeesRelation.API.Controllers
 {
@@ -11,14 +11,10 @@ namespace EmployeesRelation.API.Controllers
     [Route("[controller]")]
     public class EmployeeController : ControllerBase
     {
-
-        private readonly ILogger<EmployeeController> _logger;
-
-
-        public EmployeeController(ILogger<EmployeeController> logger)
+        private readonly Logs _logger;
+        public EmployeeController(Logs logger)
         {
             _logger = logger;
-
         }
 
         [HttpGet]
@@ -46,9 +42,17 @@ namespace EmployeesRelation.API.Controllers
         {
             List<Employee> database = JsonOperations.Read();
             var result = database.Where(x => x.Id == id).FirstOrDefault();
-            return Ok(result);
-        }
+            if (result is null)
+            {
 
+                return StatusCode(StatusCodes.Status404NotFound);
+            }
+            else
+            {
+                return Ok(result);
+            }
+
+        }
 
         [HttpPost]
         [ProducesResponseType(typeof(Employee), StatusCodes.Status201Created)]
@@ -67,7 +71,8 @@ namespace EmployeesRelation.API.Controllers
                 Salary = request.Salary,
 
             };
-
+            //_logger.Before = employee.ToString();
+            //_logger.After = null;
             JsonOperations._database.Add(employee);
             JsonOperations.Save();
             return Created(string.Empty, employee);
@@ -76,11 +81,11 @@ namespace EmployeesRelation.API.Controllers
         [ProducesResponseType(typeof(Employee), StatusCodes.Status201Created)]
         [ProducesResponseType(204)]
         [ProducesResponseType(typeof(Employee), StatusCodes.Status404NotFound)]
-        public async Task<IActionResult> GetByFilter([FromQuery] int page, [FromQuery] int maxResults, [FromBody] EmployeeParameters request)
+        public async Task<IActionResult> GetByFilter([FromQuery] int page, [FromQuery] int maxResults, [FromBody] EmployeeParameters request)//Passar Body
         {
             List<Employee> database = JsonOperations.Read();
             List<Employee> filteredData;
-            if (request.Gender == "string" && request.JobTitle != "string")
+            if (request.Gender.Equals("string") && !request.JobTitle.Equals("string"))
             {
                 filteredData = database
                                    .Where(x => x.JobTitle.Contains(request.JobTitle) && x.Salary >= request.Salary)
@@ -89,7 +94,7 @@ namespace EmployeesRelation.API.Controllers
                                    .OrderBy(x => x.Id)
                                    .ToList();
             }
-            else if (request.JobTitle == "string" && request.Gender != "string")
+            else if (request.JobTitle.Equals("string") && !request.Gender.Equals("string"))
             {
 
                 filteredData = database
@@ -99,7 +104,7 @@ namespace EmployeesRelation.API.Controllers
                                        .OrderBy(x => x.Id)
                                        .ToList();
             }
-            else if (request.JobTitle == "string" && request.Gender == "string")
+            else if (request.JobTitle.Equals("string") && request.Gender.Equals("string"))
             {
                 filteredData = database
                                        .Where(x => x.Salary >= request.Salary)
@@ -122,6 +127,7 @@ namespace EmployeesRelation.API.Controllers
         }
 
         [HttpPut("{id}")]
+        [CustomActionFilter()]
         [Consumes(MediaTypeNames.Application.Json)]
         [ProducesResponseType(typeof(Employee), StatusCodes.Status201Created)]
         [ProducesResponseType(typeof(Employee), StatusCodes.Status200OK)]
@@ -135,6 +141,7 @@ namespace EmployeesRelation.API.Controllers
             var found = database.Where(x => x.Id == id).FirstOrDefault();
             JsonOperations._database.Remove(found);
 
+
             var inserted = new Employee
             {
                 Id = id,
@@ -145,27 +152,72 @@ namespace EmployeesRelation.API.Controllers
                 Salary = request.Salary,
             };
             JsonOperations._database.Add(inserted);
+            _logger.EmployeeName = (inserted.FirstName.ToString() + " " + inserted.LastName.ToString());
+            _logger.EmployeeId = id;
+            _logger.Before = found.ToString();
+            _logger.After = inserted.ToString();
             JsonOperations.Save();
             return Ok(inserted);
         }
 
-
         [HttpPatch("{id}")]
-        public IActionResult Patch()
+        [CustomActionFilter()]
+        [ProducesResponseType(typeof(Employee), StatusCodes.Status200OK)]
+        [ProducesResponseType(typeof(Employee), StatusCodes.Status415UnsupportedMediaType)]
+        [ProducesResponseType(typeof(Employee), StatusCodes.Status404NotFound)]
+        public IActionResult Patch([FromRoute] int id, EmployeeParameters request)
         {
-            return Ok();
+            List<Employee> database = JsonOperations.Read();
+
+            var found = database.Where(x => x.Id == id).FirstOrDefault();
+
+            if (found is null)
+            {
+                return StatusCode(StatusCodes.Status404NotFound);
+            }
+            JsonOperations._database.Remove(found);
+
+            var inserted = new Employee
+            {
+                Id = id,
+                FirstName = found.FirstName,
+                LastName = found.LastName,
+                Gender = request.Gender,
+                JobTitle = request.JobTitle,
+                Salary = request.Salary,
+            };
+            JsonOperations._database.Add(inserted);
+            JsonOperations.Save();
+            _logger.EmployeeName = (inserted.FirstName.ToString() + " " + inserted.LastName.ToString());
+            _logger.EmployeeId = id;
+            _logger.Before = found.ToString();
+            _logger.After = inserted.ToString();
+            return Ok(inserted);
         }
 
         [HttpDelete("{id}")]
+        [CustomActionFilter()]
         [ProducesResponseType(typeof(Employee), 200)]
         [ProducesResponseType(204)]
+        [ProducesResponseType(typeof(Employee), StatusCodes.Status404NotFound)]
         public IActionResult Delete(int id)
         {
             List<Employee> database = JsonOperations.Read();
             var result = database.Where(x => x.Id == id).FirstOrDefault();
-            JsonOperations._database.Remove(result);
-            JsonOperations.Save();
-            return Ok("Employee Deleted");
+            if (result is null)
+            {
+                return StatusCode(StatusCodes.Status404NotFound);
+            }
+            else
+            {
+                _logger.EmployeeName = (result.FirstName.ToString() + " " + result.LastName.ToString());
+                _logger.EmployeeId = result.Id;
+                _logger.Before = result.ToString();
+                _logger.After = null;
+                JsonOperations._database.Remove(result);
+                JsonOperations.Save();
+                return Ok(result);
+            }
         }
     }
 }
